🧹 refactor: move repetative logic into helper functions
- performing AppArm pattern matching and checking payload builders
  moved into try_get_params function
- unwrapping the params into the correct params for the form moved into
  three specialized functions
- moved input plugin functions into their own modules
